test(parameters): colocate resource trigger tests with implementation
